+++ shared/core/src/main/kotlin/CheckProjectType.kt
@@ -4,11 +4,17 @@ package com.javiersc.gradle.plugins.core
 
 import org.gradle.api.Project
 
+val Project.isAndroidApplication: Boolean
+    get() = pluginManager.hasPlugin("com.android.application")
+
 val Project.isAndroidLibrary: Boolean
     get() = pluginManager.hasPlugin("com.android.library") && isKotlinMultiplatform.not()
 
 val Project.isGradlePlugin: Boolean
     get() = pluginManager.hasPlugin("org.gradle.java-gradle-plugin")
+
+val Project.isJavaPlatform: Boolean
+    get() = pluginManager.hasPlugin("org.gradle.java-platform")
 
 val Project.isKotlinJvm: Boolean
     get() = pluginManager.hasPlugin("org.jetbrains.kotlin.jvm")